requirements.txt
absl-py==0.8.0
astor==0.8.0
Click==7.0
cycler==0.10.0
Flask==0.12.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.2
numpy==1.17.1
pkg-resources==0.0.0
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
scipy==1.1.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.23.4
Werkzeug==0.15.6
wrapt==1.11.2
